audio sample packet: always send 4 samples
Some televisions seem to dislike the frequent sending of audio sample
packets with few samples (1 or 2) and will completely ignore the audio
that's being sent without any indication of an error.

This commit should deal with these TVs that are not fully compliant and
expect us to send >= 3 samples per packet, by always sending 4 samples.

I was using a set of load and clear registers, but now I've adopted a
buffer-swapping technique similar to double-buffering in video.