hdmi: allow setting start x/y with parameters
gbaHD needs this synchronize the hdmi output with the start of a GBA
frame, with the GBA image centered within the HDMI frame instead of
starting at 0,0.

It's easier and uses fewer FPGA resources to start the hdmi block in the
middle of a frame than to count off the right number of cycles from GBA
vsync to achieve the same phase relationship.  Receivers need to be able
to synchronize from mid-stream anyway, so this doesn't cause any
problems.